refactor(generate): restructure
