+++ packages/generate/models/openai.ts
